execenv: move terminal detection to Out, introduce the compagnion In
