+++ commands/execenv/env_testing.go
@@ -13,10 +13,26 @@ import (
 	"github.com/MichaelMure/git-bug/repository"
 )
 
+var _ In = &TestIn{}
+
+type TestIn struct {
+	*bytes.Buffer
+	forceIsTerminal bool
+}
+
+func (t *TestIn) ForceIsTerminal(value bool) {
+	t.forceIsTerminal = value
+}
+
+func (t *TestIn) IsTerminal() bool {
+	return t.forceIsTerminal
+}
+
 var _ Out = &TestOut{}
 
 type TestOut struct {
 	*bytes.Buffer
+	forceIsTerminal bool
 }
 
 func (te *TestOut) Printf(format string, a ...interface{}) {
@@ -40,13 +56,29 @@ func (te *TestOut) PrintJSON(v interface{}) error {
 	return nil
 }
 
+func (te *TestOut) IsTerminal() bool {
+	return te.forceIsTerminal
+}
+
 func (te *TestOut) Raw() io.Writer {
 	return te.Buffer
 }
 
+func (te *TestOut) ForceIsTerminal(value bool) {
+	te.forceIsTerminal = value
+}
+
 func NewTestEnv(t *testing.T) *Env {
 	t.Helper()
+	return newTestEnv(t, false)
+}
 
+func NewTestEnvTerminal(t *testing.T) *Env {
+	t.Helper()
+	return newTestEnv(t, true)
+}
+
+func newTestEnv(t *testing.T, isTerminal bool) *Env {
 	repo := repository.CreateGoGitTestRepo(t, false)
 
 	backend, err := cache.NewRepoCacheNoEvents(repo)
@@ -59,7 +91,8 @@ func NewTestEnv(t *testing.T) *Env {
 	return &Env{
 		Repo:    repo,
 		Backend: backend,
-		Out:     &TestOut{&bytes.Buffer{}},
-		Err:     &TestOut{&bytes.Buffer{}},
+		In:      &TestIn{Buffer: &bytes.Buffer{}, forceIsTerminal: isTerminal},
+		Out:     &TestOut{Buffer: &bytes.Buffer{}, forceIsTerminal: isTerminal},
+		Err:     &TestOut{Buffer: &bytes.Buffer{}, forceIsTerminal: isTerminal},
 	}
 }
